2170 - Added Channel name fallback
